refactor: remove DHIS server version filtering
