+++ src/components/AppDetails.module.css
