+++ src/components/AppDetails/channel-to-display-name.js
